Moved POJO
Moved Hashtag POJO from its own package to the base package where all other POJOs are. Deleted hashtagpojo package. HashtagStorage refactored imports
