+++ src/main/java/org/wecancodeit/reviews/Hashtag.java
@@ -1,4 +1,4 @@
-package hashtagpojo;
+package org.wecancodeit.reviews;
 
 public class Hashtag {
